requirements.txt
Flask~=2.2.0
Flask-RESTful~=0.3.9
Flask-Cors~=3.0.10
Flask-Login~=0.6.2
flask-marshmallow~=0.14.0
Flask-Security-Too~=5.1.1
Flask-SQLAlchemy~=3.0.3
marshmallow~=3.19.0
pip~=23.0.1
argparse~=1.4.0
wheel~=0.37.1
bcrypt~=4.0.1
cryptography~=3.4.8
keyring~=23.5.0
setuptools~=66.1.1
future~=0.18.3
matplotlib~=3.6.3
numpy~=1.24.1
python-dateutil~=2.8.2
celery[redis]~=5.2.7
redis~=4.5.2
secure-smtplib~=0.1.1
lxml~=4.9.2
Jinja2~=3.1.2
Werkzeug~=2.2.3
Brotli~=1.0.9
weasyprint~=58.1
Flask-Caching~=2.0.2
Pillow~=9.4.0
pandas~=1.5.3